First commit
The difficult part (aka God-forsaken sed) seems to be done. Still need to test with actual server syncing and make sure that on the event that one rsync fails, on the next start we're able to get both parties synced (probably a two-way non-destructive rsync with --update). So far this version is built on the very bad assumption that wifi somehow doesn't suck